requirements.txt
aiofiles==25.1.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
coverage==7.13.0
cryptography==46.0.3
dotenv==0.9.9
greenlet==3.3.0
h11==0.16.0
hiredis==3.3.0
html5tagger==1.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
marshmallow==3.26.1
multidict==6.7.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
portalocker==3.2.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sanic==25.3.0
Sanic-Cors==2.2.0
sanic-ext==23.12.0
sanic-openapi==21.12.0
sanic-redis==0.6.0
sanic-routing==23.12.0
sanic-testing==24.6.0
setuptools==80.9.0
SQLAlchemy==2.0.45
tracerite==1.1.3
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.2
uvloop==0.22.1
websockets==15.0.1
